Move old package out of the way
