+++ src/go/cmdv1/doc.go
@@ -1,4 +1,4 @@
 // The Command Line Interface (CLI) for Marmot.
 //   - Contains: a Command Line Interface and the factories it uses to make services and use cases.
 //   - Dependencies: all other layers.
-package cmd
+package cmdv1
