+++ src/go/cmdv1/init.go
@@ -1,10 +1,10 @@
-package cmd
+package cmdv1
 
 import (
 	"fmt"
 	"io"
 
-	cmdroot "github.com/kkrull/marmot/cmd/root"
+	cmdroot "github.com/kkrull/marmot/cmdv1/root"
 	"github.com/spf13/cobra"
 )
