+++ src/go/cmdv1/local/register.go
@@ -1,7 +1,7 @@
 package cmdlocal
 
 import (
-	cmdroot "github.com/kkrull/marmot/cmd/root"
+	cmdroot "github.com/kkrull/marmot/cmdv1/root"
 	"github.com/spf13/cobra"
 )
