+++ src/go/cmdv1/remote/list.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"io"
 
-	cmdroot "github.com/kkrull/marmot/cmd/root"
+	cmdroot "github.com/kkrull/marmot/cmdv1/root"
 	"github.com/spf13/cobra"
 )
