+++ src/go/cmdv1/remote/register.go
@@ -5,7 +5,7 @@ import (
 	"net/url"
 	"strings"
 
-	cmdroot "github.com/kkrull/marmot/cmd/root"
+	cmdroot "github.com/kkrull/marmot/cmdv1/root"
 	"github.com/spf13/cobra"
 )
