+++ src/go/cmdv1/root_groups.go
@@ -1,4 +1,4 @@
-package cmd
+package cmdv1
 
 import (
 	"fmt"
